Move forced variations into DecisionService.
