+++ src/Optimizely/ProjectConfig.php
@@ -140,14 +140,6 @@ class ProjectConfig
     private $_errorHandler;
 
     /**
-     * @var array Associative array of user IDs to an associative array
-     * of experiments to variations. This contains all the forced variations
-     * set by the user by calling setForcedVariation (it is not the same as the
-     * whitelisting forcedVariations data structure in the Experiments class).
-     */
-    private $_forcedVariationMap;
-
-    /**
      * list of Feature Flags that will be parsed from the datafile.
      *
      * @var [FeatureFlag]
@@ -214,7 +206,6 @@ class ProjectConfig
         $this->_anonymizeIP = isset($config['anonymizeIP'])? $config['anonymizeIP'] : false;
         $this->_botFiltering = isset($config['botFiltering'])? $config['botFiltering'] : null;
         $this->_revision = $config['revision'];
-        $this->_forcedVariationMap = [];
 
         $groups = $config['groups'] ?: [];
         $experiments = $config['experiments'] ?: [];
@@ -607,93 +598,7 @@ class ProjectConfig
         return isset($this->_variationIdMap[$experimentKey]) &&
             isset($this->_variationIdMap[$experimentKey][$variationId]);
     }
-
-    /**
-     * Gets the forced variation key for the given user and experiment.
-     *
-     * @param $experimentKey string Key for experiment.
-     * @param $userId string The user Id.
-     *
-     * @return Variation The variation which the given user and experiment should be forced into.
-     */
-    public function getForcedVariation($experimentKey, $userId)
-    {
-
-        if (!isset($this->_forcedVariationMap[$userId])) {
-            $this->_logger->log(Logger::DEBUG, sprintf('User "%s" is not in the forced variation map.', $userId));
-            return null;
-        }
-
-        $experimentToVariationMap = $this->_forcedVariationMap[$userId];
-        $experimentId = $this->getExperimentFromKey($experimentKey)->getId();
-        // check for null and empty string experiment ID
-        if (strlen($experimentId) == 0) {
-            // this case is logged in getExperimentFromKey
-            return null;
-        }
-
-        if (!isset($experimentToVariationMap[$experimentId])) {
-            $this->_logger->log(Logger::DEBUG, sprintf('No experiment "%s" mapped to user "%s" in the forced variation map.', $experimentKey, $userId));
-            return null;
-        }
-
-        $variationId = $experimentToVariationMap[$experimentId];
-        $variation = $this->getVariationFromId($experimentKey, $variationId);
-        $variationKey = $variation->getKey();
-
-        $this->_logger->log(Logger::DEBUG, sprintf('Variation "%s" is mapped to experiment "%s" and user "%s" in the forced variation map', $variationKey, $experimentKey, $userId));
-
-        return $variation;
-    }
-
-    /**
-     * Sets an associative array of user IDs to an associative array of experiments
-     * to forced variations.
-     *
-     * @param $experimentKey string Key for experiment.
-     * @param $userId string The user Id.
-     * @param $variationKey string Key for variation. If null, then clear the existing experiment-to-variation mapping.
-     *
-     * @return boolean A boolean value that indicates if the set completed successfully.
-     */
-    public function setForcedVariation($experimentKey, $userId, $variationKey)
-    {
-
-        // check for empty string Variation key
-        if (!is_null($variationKey) && !Validator::validateNonEmptyString($variationKey)) {
-            $this->_logger->log(Logger::ERROR, sprintf(Errors::INVALID_FORMAT, Optimizely::VARIATION_KEY));
-            return false;
-        }
-
-        $experiment = $this->getExperimentFromKey($experimentKey);
-        $experimentId = $experiment->getId();
-        // check if the experiment exists in the datafile (a new experiment is returned if it is not in the datafile)
-        if (strlen($experimentId) == 0) {
-            // this case is logged in getExperimentFromKey
-            return false;
-        }
-
-        // clear the forced variation if the variation key is null
-        if (is_null($variationKey)) {
-            unset($this->_forcedVariationMap[$userId][$experimentId]);
-            $this->_logger->log(Logger::DEBUG, sprintf('Variation mapped to experiment "%s" has been removed for user "%s".', $experimentKey, $userId));
-            return true;
-        }
-
-        $variation = $this->getVariationFromKey($experimentKey, $variationKey);
-        $variationId = $variation->getId();
-        // check if the variation exists in the datafile (a new variation is returned if it is not in the datafile)
-        if (strlen($variationId) == 0) {
-            // this case is logged in getVariationFromKey
-            return false;
-        }
-
-        $this->_forcedVariationMap[$userId][$experimentId] = $variationId;
-        $this->_logger->log(Logger::DEBUG, sprintf('Set variation "%s" for experiment "%s" and user "%s" in the forced variation map.', $variationId, $experimentId, $userId));
-
-        return true;
-    }
-
+    
     /**
      * Determines if given experiment is a feature test.
      *
